test front end started

+++ src/Controller/ReservationController.php
@@ -11,6 +11,7 @@ use App\Entity\Placement;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Constraints as Assert;
@@ -22,8 +23,9 @@ class ReservationController extends AbstractController
     public function getAllReservations(Request $request,ReservationRepository $reservationRepository): JsonResponse
     {
         // now day for date end by default
-        $dateStart = $request->get('dateStart', default: date("Y-m-d", strtotime("-1 month")));
-        $dateEnd = $request->get('dateEnd', default: date('Y-m-d'));
+        $dateStart = $request->get('dateStart', default: date("Y-m-d", strtotime("now")));
+        $dateEnd = $request->get('dateEnd', default: date("Y-m-d", strtotime("+1 week")));
+
         //create a validator
         $validator = Validation::createValidator();
         //check if date range is ok
@@ -48,45 +50,39 @@ class ReservationController extends AbstractController
         $reservations = $reservationRepository->findAllByDate($dateStart, $dateEnd);
         // format reservations per day and return result
         return $this->json($reservationRepository->groupByDay($reservations));
-        // format reservations per day and return result
     }
 
-    #[Route('/add/{date}/{placement}/{foodtruck}',
-        name: 'reservation.add',
-        requirements: [
-            'date'=>'\d{4}-\d{2}-\d{2}',
-            'placement'=>'\d{1}'
-        ],
-        methods: ['GET']
-    )]
-    public function addReservation(
+    #[Route('/add', name: 'reservation.add', methods: ['POST'])]
+    public function addReservation(Request $request,
                     EntityManagerInterface $entityManager,
-                    string $date,
-                    int $placement,
-                    string $foodtruck): JsonResponse
+                    #[MapRequestPayload(
+                        validationGroups:['reservation.add'] ,
+                        validationFailedStatusCode: 400)] Reservation $reservation): JsonResponse
     {
+        // get content to add from json file
+        $toadd = json_decode($request->getContent(), true);
         //create a validator
         $validator = Validation::createValidator();
         //check if date range is ok
-        $violations = $validator->validate([$date], [
-            new Assert\All([new Assert\Date()])
+        $violations = $validator->validate($toadd['date'], [
+            new Assert\Date()
         ]);
         // init errors
         $errors=[];
         // If error we display and stop execution
         if (0 !== count($violations))$errors[]='Wrong date '. $violations[0]->getInvalidValue();
         // find the related foodtruck
-        $foodtruck_obj = $entityManager->getRepository(Foodtruck::class)->find($foodtruck);
-        if(!$foodtruck_obj)$errors[]='Food truck not found';
+        if(!$foodtruck_obj = $entityManager->getRepository(Foodtruck::class)->find($toadd['foodtruck']['id']))
+            $errors[]='Foodtruck not found';
         // find the related placement
-        $placement_obj = $entityManager->getRepository(Placement::class)->find($placement);
-        if(!$placement_obj)$errors[]='placement not found';
+        if(!$placement_obj = $entityManager->getRepository(Placement::class)->find($toadd['placement']['id']))
+            $errors[]='Placement not found';
         // If errors => error 400
         if($errors)return $this->json($errors,status:400);
 
-        // set a new reservation
+        // set a new reservation (to avoid reference id issue)
         $reservation = (new Reservation())
-            ->setDate(DateTimeImmutable::createFromFormat('Y-m-d', $date))
+            ->setDate(DateTimeImmutable::createFromFormat('Y-m-d', $toadd['date']))
             ->setFoodtruck($foodtruck_obj)
             ->setPlacement($placement_obj);
         // persiste the reservation
@@ -99,16 +95,21 @@ class ReservationController extends AbstractController
         ]);
     }
 
-    #[Route('/delete/{id}',name: 'delete_product',methods:['DELETE'])]
-    public function delete(Reservation $reservation,EntityManagerInterface $entityManager): JsonResponse
+    #[Route('/delete/{id}', name: 'delete', requirements: ['placement'=>'\d+'],methods:['DELETE'])]
+    public function delete(EntityManagerInterface $entityManager,int $id): JsonResponse
     {
-        $res = $reservation;
-        // delete a product
-        $entityManager->remove($reservation);
-        $entityManager->flush();
-        return $this->json($res,200,[],[
-            'groups'=>['reservation.delete'],
-        ]);
+        // if the reservation exists we delete it
+        if($reservation = $entityManager->getRepository(Reservation::class)->find($id))
+        {
+            // delete a product
+            $entityManager->remove($reservation);
+            $entityManager->flush();
+            return $this->json($reservation,200,[],[
+                'groups'=>['reservation.delete'],
+            ]);
+        }
+        // otherwise return unfound error
+        return $this->json(['Reservation not found'],status:400);
     }
 
 }
